requirements.txt
tweepy>=3.10.0
python-socketio==4.6.1
pytz>=2021.1
cachetools>=4.2.1